added repeat recharge

+++ backend/jwt/src/main/java/com/webzy/jwt/service/AddonServiceImpl.java
@@ -2,7 +2,7 @@ package com.webzy.jwt.service;
 
 import org.springframework.stereotype.Service;
 
-import com.webzy.jwt.dao.AddonRepo;
+import com.webzy.jwt.repository.AddonRepo;
 import com.webzy.jwt.entity.Addon;
 
 import jakarta.annotation.PostConstruct;
@@ -20,38 +20,36 @@ public class AddonServiceImpl implements AddonService {
     // @PostConstruct
     public void init() {
         List<Addon> jioAddons = Arrays.asList(
-                createAddon("Extra Data", "1GB", 15, "Get 1 GB Data", "1 day", "Jio"),
-                createAddon("Roaming Pack", "1.5GB", 19, "Data for roaming usage", "1 day", "Jio"),
-                createAddon("Social Media Pack", "2GB", 25, "Data for social media apps", "1 day", "Jio"),
-                createAddon("Video Streaming Pack", "2.5GB", 29, "Data for video streaming apps", "15 days", "Jio"),
-                createAddon("Gaming Pack", "6GB", 61, "Data for gaming apps", "10 days", "Jio"),
-                createAddon("Business Pack", "12GB", 121, "Data for business applications", "24 days", "Jio"),
-                createAddon("Night Pack", "10GB", 148, "Unlimited data for nighttime usage", "28 days", "Jio"),
-                createAddon("Weekend Pack", "30GB", 181, "Extra data for weekends", "30 days", "Jio"),
-                createAddon("Music Streaming Pack", "50GB", 301, "Get 50 GB Data for all Customers.", "7 days", "Jio"),
-                createAddon("Family Pack", "40GB", 241, "Get 40 GB Data for all Customers.", "30 days", "Jio"),
-                //Airtel
-                createAddon("Extra Data", "1GB", 15, "Additional 2GB data", "1 day", "Airtel"),
-                createAddon("Roaming Pack", "1.5GB", 20, "Data for roaming usage", "1 day", "Airtel"),
-                createAddon("Social Media Pack", "2GB", 25, "Data for social media apps", "1 day", "Airtel"),
-                createAddon("Video Streaming Pack", "2.5GB", 30, "Data for video streaming apps", "15 days", "Airtel"),
-                createAddon("Gaming Pack", "6GB", 70, "Data for gaming apps", "10 days", "Airtel"),
-                createAddon("Business Pack", "12GB", 120, "Data for business applications", "24 days", "Airtel"),
-                createAddon("Night Pack", "10GB", 150, "Unlimited data for nighttime usage", "28 days", "Airtel"),
-                createAddon("Weekend Pack", "30GB", 179, "Extra data for weekends", "30 days", "Airtel"),
-                createAddon("Music Streaming Pack", "50GB", 300, "Data for music streaming apps", "7 days", "Airtel"),
-                createAddon("Family Pack", "50GB", 250, "Shared data for family members", "30 days", "Airtel"));
+                createAddon("Extra Data", "1GB", 15, "Get 1 GB Data", "Jio"),
+                createAddon("Roaming Pack", "1.5GB", 19, "Data for roaming usage", "Jio"),
+                createAddon("Social Media Pack", "2GB", 25, "Data for social media apps", "Jio"),
+                createAddon("Video Streaming Pack", "2.5GB", 29, "Data for video streaming apps", "Jio"),
+                createAddon("Gaming Pack", "6GB", 61, "Data for gaming apps", "Jio"),
+                createAddon("Business Pack", "12GB", 121, "Data for business applications", "Jio"),
+                createAddon("Night Pack", "10GB", 148, "Unlimited data for nighttime usage", "Jio"),
+                createAddon("Weekend Pack", "30GB", 181, "Extra data for weekends", "Jio"),
+                createAddon("Music Streaming Pack", "50GB", 301, "Get 50 GB Data for all Customers.", "Jio"),
+                createAddon("Family Pack", "40GB", 241, "Get 40 GB Data for all Customers.", "Jio"),
+                // Airtel
+                createAddon("Extra Data", "1GB", 15, "Additional 2GB data", "Airtel"),
+                createAddon("Roaming Pack", "1.5GB", 20, "Data for roaming usage", "Airtel"),
+                createAddon("Social Media Pack", "2GB", 25, "Data for social media apps", "Airtel"),
+                createAddon("Video Streaming Pack", "2.5GB", 30, "Data for video streaming apps", "Airtel"),
+                createAddon("Gaming Pack", "6GB", 70, "Data for gaming apps", "Airtel"),
+                createAddon("Business Pack", "12GB", 120, "Data for business applications", "Airtel"),
+                createAddon("Night Pack", "10GB", 150, "Unlimited data for nighttime usage", "Airtel"),
+                createAddon("Weekend Pack", "30GB", 179, "Extra data for weekends", "Airtel"),
+                createAddon("Music Streaming Pack", "50GB", 300, "Data for music streaming apps", "Airtel"),
+                createAddon("Family Pack", "50GB", 250, "Shared data for family members", "Airtel"));
         addonRepository.saveAll(jioAddons);
     }
 
-    private Addon createAddon(String addonName, String data, double addonPrice, String addonDetails,
-            String addonValidity, String operatorName) {
+    private Addon createAddon(String addonName, String data, double addonPrice, String addonDetails, String operatorName) {
         Addon addon = new Addon();
         addon.setAddonName(addonName);
         addon.setData(data);
         addon.setAddonPrice(addonPrice);
         addon.setAddonDetails(addonDetails);
-        addon.setAddonValidity(addonValidity);
         addon.setOperatorName(operatorName);
         return addon;
     }
@@ -67,8 +65,8 @@ public class AddonServiceImpl implements AddonService {
     }
 
     @Override
-    public Addon createAddon(Addon addon) {
-        return addonRepository.save(addon);
+    public boolean createAddon(Addon addon) {
+        return addonRepository.save(addon) != null ? true : false;
     }
 
     @Override
@@ -78,12 +76,16 @@ public class AddonServiceImpl implements AddonService {
     }
 
     @Override
-    public void deleteAddon(Long id) {
-        addonRepository.deleteById(id);
+    public boolean deleteAddon(Long id) {
+        if (addonRepository.findById(id).isPresent()) {
+            addonRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public List<Addon> getAddOnByOperatorName(String operatorName){
+    public List<Addon> getAddOnByOperatorName(String operatorName) {
         return addonRepository.findByOperatorName(operatorName);
     }
 
@@ -104,9 +106,6 @@ public class AddonServiceImpl implements AddonService {
         }
         if (updatedAddon.getAddonDetails() != null) {
             existingAddon.setAddonDetails(updatedAddon.getAddonDetails());
-        }
-        if (updatedAddon.getAddonValidity() != null) {
-            existingAddon.setAddonValidity(updatedAddon.getAddonValidity());
         }
         if (updatedAddon.getOperatorName() != null) {
             existingAddon.setOperatorName(updatedAddon.getOperatorName());
